requirements.txt
Django>=4.2.7,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.1.0
pytz>=2023.3
Pillow>=9.5.0
python-dotenv>=1.0.0
gunicorn>=21.0.0
whitenoise>=6.5.0
dj-database-url>=2.1.0
psycopg2-binary>=2.9.7  # For PostgreSQL support
# mysqlclient>=2.2.0  # For MySQL support (optional) - Removed as not needed
redis>=5.0.0  # For caching and session storage
django-redis>=5.3.0
django-storages>=1.14.0  # For cloud storage (optional)
django-debug-toolbar>=4.2.0  # For development only
